Add basic image fetching object
This object is pretty naive and could use some love. It attempts to
avoid threading issues using a private queue that is used for all
reads and writes to the underlying URL -> data task map. This map this
is what holds the reference to any created download task for an image.
The rendering of the UIImage does happen off the main thread so that
there is as little stutter as possible in the UI.