Use ImageFetcher to get posters in results
Now the image fetching mechanism is passed down to the results view
controller where it is used. The results controller ensures that once
the image is provided that the cell it is applied to is the one
displaying the same movie. If there is no cell at the index path then
the image is not set.

This is a bit messy but without using a more robust and battle tested
library it is moderately reasonable.

Another approach would have been even further passing of the
ImageFetcher down into the individual cells which then interally use
that to set the poster image via a URL property that is also exposed.