Add a movie detail view controller
This essentially chains the fetch for the poster image *after* the
details have been queried. This ordering isn't absolutely necessary
since the search result already had the poster URL. Of course there is
no guarantee that the API will return the same URL for both the search
and the detail view. Further, the design for this detail view controller
is pretty rough, but is a first pass.